requirements.txt
-e git+https://github.com/cool-soft/boiler-softm.git#egg=boiler-softm
-e git+https://github.com/cool-soft/boiler.git#egg=boiler

pandas~=1.2.4
python-dateutil~=2.8.1

numpy~=1.18.5
tensorflow~=2.3.0